Add S3 user metadata to uploads.
S3 uploads can be tagged with user metadata in the form of key-value
pairs. Add a config option on a project-by-project basis to tag uploads
with this metadata.

This is most useful when the metadata includes Jenkins macros–for
instance, we plan to use it to tag build uploads with the branch of the
build.

The option is project-by-project rather than file-by-file due to
Jenkins/Jelly issues with nested <f:repeatable> elements.