[FIXED JENKINS-14470]
The instance field needs to be marked as transient to avoid persistence. In this case, I see no point in AtomicReference since a race condition of creating two clients will do no real harm.

So I just replaced it with volatile field to avoid the need of recreating AtomicReference object upon the deserialization from XStream.