S3BucketPublisher: handle InterruptedException
Previously, if the source file patterns given to the S3 publisher
matched no files, we tried to be nice and called validateAntFileMask(),
which gave a nicely formatted message in case some other pattern would
have matched.

However, if even validateAntFileMask() fails to match anything within
the bound limit, then we would get a big fat InterruptedException that
kills the whole publisher and marks the build as a failure.

This patch makes this consistent by just ignoring the exception if
nothing was found. IOW, if we have nothing to suggest to the user to
help them diagnose the issue, that shouldn't be a reason to die.