JENKINS-27284: Constrain build result severity
This change adds a per-job setting to constrain the build result
nastiness set by the plugin for problems that occur inside the plugin.
That is, an S3 failure that would normally result in a FAILURE can
be constrained to UNSTABLE or SUCCESS.

This addresses JENKINS-27284
<https://issues.jenkins-ci.org/browse/JENKINS-27284> by providing a
way to prevent S3 plugin failures from causing an otherwise successful
build to be marked as having failed.