S3BaseUploadCallable: add missing break for 'content-type'
When handling the metadata for the object getting pushed to S3, the
special case of 'content-type' appears to be missing a break.  This
causes the object to have the 'Content-Type' header and the
'x-amz-meta-content-type' header.

While this doesn't seem to break anything, I believe it is the correct
way to do this.